+++ backend/src/BMKPlace.Domain/Exceptions/UserCooldownException.cs
@@ -5,11 +5,11 @@ namespace BMKPlace.Domain.Exceptions;
 
 public class UserCooldownException : DomainValidationException
 {
-    public long UserId { get; }
+    public int UserId { get; }
     public int CanvasId { get; }
     public DateTimeOffset CooldownEndTime { get; }
 
-    public UserCooldownException(long userId, int canvasId, DateTimeOffset cooldownEndTime)
+    public UserCooldownException(int userId, int canvasId, DateTimeOffset cooldownEndTime)
         : base($"User {userId} on canvas {canvasId} is on cooldown until {cooldownEndTime:o}.")
     {
         UserId = userId;
